Move: ClosureComponent 파일 분리
- Components 폴더 추가
- ClosureComponent 파일 분리
