+++ APPRO/APPRO/Modules/HeadTracker.swift
@@ -40,30 +40,3 @@ final class HeadTracker: ObservableObject {
     }
     
 }
-
-
-struct ClosureComponent: Component {
-    
-    let closure: (TimeInterval) -> Void
-    
-    init(closure: @escaping (TimeInterval) -> Void) {
-        self.closure = closure
-    }
-    
-}
-
-struct ClosureSystem: System {
-    
-    static let query = EntityQuery(where: .has(ClosureComponent.self))
-    
-    init(scene: Scene) {}
-    
-    func update(context: SceneUpdateContext) {
-        for entity in context.entities(matching: Self.query, updatingSystemWhen: .rendering) {
-            guard let component = entity.components[ClosureComponent.self] else { return }
-            
-            component.closure(context.deltaTime)
-        }
-    }
-    
-}
